+++ lib/arel_extensions/visitors/oracle.rb
@@ -388,6 +388,8 @@ module ArelExtensions
       end
 
       def visit_ArelExtensions_Nodes_Substring(o, collector)
+        return visit(o.range_substring_node, collector) if o.range?
+
         collector << 'SUBSTR('
         o.expressions.each_with_index { |arg, i|
           collector << COMMA if i != 0